Initial version of Basket page implemented with API consumption.

+++ src/project/SRP.Application/Features/Baskets/Commands/Add/BasketAddCommand.cs
@@ -7,7 +7,7 @@ namespace SRP.Application.Features.Baskets.Commands.Add;
 public class BasketAddCommand : IRequest<string>, ITransactional, ILoggableRequest
 {
     public decimal Price { get; set; }
-    public decimal Count { get; set; }
+    public int Count { get; set; }
     public decimal TotalPrice { get; set; }
     public int ProductID { get; set; }
     public int MenuTableID { get; set; }
